backend/generated_projects/f477a7bf-0229-4c69-a70e-ec75648f73c6/backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
pymongo==4.6.0
python-multipart==0.0.6
email-validator==2.1.0